Refactor capacity fields from Long to int in event, venue, and cart item models

+++ backend/cart-service/src/main/java/me/doruk/cartservice/response/CatalogServiceResponse.java
@@ -14,7 +14,7 @@ import java.math.BigDecimal;
 public class CatalogServiceResponse {
   private Long eventId;
   private String event;
-  private Long capacity;
+  private int capacity;
   private VenueResponse venue;
   private BigDecimal ticketPrice;
 }
