+++ backend/cart-service/src/main/java/me/doruk/cartservice/response/VenueResponse.java
@@ -13,5 +13,5 @@ public class VenueResponse {
   private Long id;
   private String name;
   private String address;
-  private Long totalCapacity;
+  private int totalCapacity;
 }
