+++ backend/catalog-service/src/main/java/me/doruk/catalogservice/request/VenueCreateRequest.java
@@ -12,5 +12,5 @@ import lombok.NoArgsConstructor;
 public class VenueCreateRequest {
   private String name;
   private String address;
-  private Long totalCapacity;
+  private int totalCapacity;
 }
